Removing reference for custom init script on Mongo
It wasn't possible to implement it practically on Helm because the
«initdbScriptsConfigMap» parameter for MongoDB chart can be set
only on Values.yaml, rendering it impossible to template. Thus,
the ConfigMap can only have a static name. Therefore, it would be
impossible to deploy multiple Helm releases of RC to the same
namespace. Losing such capability not only breaks the CI on the
relase Chart as do not solve any significative issue. To put it
bluntly #51 isn't a deal breaker bug, so we most probably won't
fix it for now.